+++ apps/code-infra-dashboard/src/components/DailyBenchmarkChart.tsx
@@ -12,17 +12,7 @@ import type { BenchmarkReport, BenchmarkReportEntry } from '@/lib/benchmark/type
 import { useDailyCommits, GitHubCommit } from '../hooks/useDailyCommits';
 import { useCiReports } from '../hooks/useCiReports';
 import ErrorDisplay from './ErrorDisplay';
-
-const CHART_COLORS = [
-  '#1976d2', // Blue
-  '#d32f2f', // Red
-  '#2e7d32', // Green
-  '#ed6c02', // Orange
-  '#9c27b0', // Purple
-  '#00796b', // Teal
-  '#f57c00', // Amber
-  '#5d4037', // Brown
-];
+import { CHART_COLORS } from './chartColors';
 
 const durationFormatter = new Intl.NumberFormat(undefined, {
   minimumFractionDigits: 2,
