+++ apps/code-infra-dashboard/src/components/chartColors.ts
@@ -1,0 +1,10 @@
+export const CHART_COLORS = [
+  '#1976d2', // Blue
+  '#d32f2f', // Red
+  '#2e7d32', // Green
+  '#ed6c02', // Orange
+  '#9c27b0', // Purple
+  '#00796b', // Teal
+  '#f57c00', // Amber
+  '#5d4037', // Brown
+];
