+++ apps/code-infra-dashboard/src/lib/bundleSize/types.ts
@@ -1,2 +1,10 @@
 export type SizeSnapshotEntry = { parsed: number; gzip: number };
 export type SizeSnapshot = Record<string, SizeSnapshotEntry>;
+
+export interface SizeSnapshotMetadata {
+  trackedBundles?: string[];
+}
+
+export type SizeSnapshotWithMetadata = SizeSnapshot & {
+  _metadata?: SizeSnapshotMetadata;
+};
